api/pyproject.toml
[project]
name = "parcelbox"
version = "1.0.0"
description = "ParcelBox backend API server"
authors = [
    { name = "fslaktern" }
]
requires-python = ">=3.9"
dependencies = [
    "fastapi[standard]",
    "pycryptodome"
]

[project.scripts]
api = "src.parcelbox.main:main"

[build-system]
requires = ["uv_build>=0.8.5,<0.9.0"]
build-backend = "uv_build"
